Calendar - implement non-round hours
allow events to have any time value. the event will be rounded to the nearest 5 minutes and displayed on the calendar.
Before this work, events could only be on the hour (e.g. 11:00 - 19:00), and and exact number of hours as well.

5 minute periods seems like a reasonable degree of accuracy for a shift to be displayed on the calendar. Each hour is split into 12 sections. As each hour is not that big, theses sections are pretty small so there is not really the need to split them futher.

Futher performance and UX testing could be done to determine whether splitting the grid lines into e.g. minutes would have a negative impact on performance, and would be noticeable on the calendar vs 5 minute increments.