Perf: Improve pagination rendering speed
Previous versions had a 600ms~ delay when paginating.
Now the delay is hardly noticeable
This was caused by displaying an element on the screen for every time block, for every day. e.g. 48 time blocks per day * 7 days = 336 elements to position and render
Now, we render one element across the whole calendar week for horizontal lines, and one element for each whole day for the vertical lines, so we only render 48 + 7 = elements.